# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flame Sales & Expense Tracker API",
    "version": "1.0.0",
    "description": "API for managing organizations, projects, expenses, sales, and related resources in the Flame Sales & Expense Tracker."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://flame-expense-tracker.bagumajonah3.workers.dev",
      "description": "Production server (Cloudflare Workers)"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Flame API key in the format flame_ak_<prefix>_<secret>. Include it as an X-Api-Key header when calling these endpoints."
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/invoices/generate": {
      "post": {
        "operationId": "generateInvoiceFromSales",
        "tags": [
          "Invoices"
        ],
        "summary": "Generate an invoice PDF from existing sales",
        "description": "Aggregates multiple sales for a single customer into a single invoice and returns a PDF.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "saleIds": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "recipient": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "address": {
                        "type": "object",
                        "nullable": true,
                        "properties": {
                          "street": {
                            "type": "string",
                            "nullable": true
                          },
                          "postalCode": {
                            "type": "string",
                            "nullable": true
                          },
                          "city": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "required": [
                      "name"
                    ]
                  },
                  "invoiceNumber": {
                    "type": "string",
                    "nullable": true
                  },
                  "invoiceDate": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  },
                  "servicePeriod": {
                    "type": "string",
                    "nullable": true
                  },
                  "currency": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "saleIds",
                  "recipient"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice PDF generated successfully.",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "API key required."
          }
        }
      }
    },
...
  }
}